@@ -4,10 +4,26 @@
  */
 package com.expressswallows.model.menu.factories;
 
+import com.expressswallows.model.menu.fooditems.*;
+import com.expressswallows.model.menu.fooditems.burgers.Burger;
+import com.expressswallows.model.menu.fooditems.drinks.Drink;
+import com.expressswallows.model.menu.fooditems.hotdogs.Hotdog;
+import com.expressswallows.model.menu.fooditems.pizzas.Pizza;
+
 /**
  *
  * @author shahi
  */
 public abstract class AbstractFactory {
-    
+
+    public abstract Burger createBurger(String request);
+
+    public abstract Pizza createPizza(String request);
+
+    public abstract Drink createDrink(String request, String size);
+
+    public abstract Hotdog createHotdog(String request);
+
+    public abstract Fries createFries(String request);
+
 }
