@@ -4,10 +4,49 @@
  */
 package com.expressswallows.model.menu.factories;
 
+import com.expressswallows.model.menu.fooditems.Fries;
+import com.expressswallows.model.menu.fooditems.burgers.Burger;
+import com.expressswallows.model.menu.fooditems.drinks.Drink;
+import com.expressswallows.model.menu.fooditems.hotdogs.Hotdog;
+import com.expressswallows.model.menu.fooditems.pizzas.Pizza;
+
 /**
  *
  * @author shahi
  */
-public class FriesFactory {
-    
+public class FriesFactory extends AbstractFactory {
+
+
+    @Override
+    public Burger createBurger(String request) {
+        return null;
+    }
+
+    @Override
+    public Pizza createPizza(String request) {
+        return null;
+    }
+
+    @Override
+    public Drink createDrink(String request, String size) {
+        return null;
+    }
+
+    @Override
+    public Hotdog createHotdog(String request) {
+        return null;
+    }
+
+    @Override
+    public Fries createFries(String request) {
+        if (request.equals("Small")) {
+            return new Fries(Fries.FriesSize.SMALL);
+        } else if (request.equals("Medium")) {
+            return new Fries(Fries.FriesSize.MEDIUM);
+        } else if (request.equals("Large")) {
+            return new Fries(Fries.FriesSize.LARGE);
+        }
+        return null;
+    }
+
 }
