@@ -4,10 +4,51 @@
  */
 package com.expressswallows.model.menu.factories;
 
+import com.expressswallows.model.menu.fooditems.Fries;
+import com.expressswallows.model.menu.fooditems.burgers.Burger;
+import com.expressswallows.model.menu.fooditems.drinks.Drink;
+import com.expressswallows.model.menu.fooditems.hotdogs.Hotdog;
+import com.expressswallows.model.menu.fooditems.pizzas.AllDressedPizza;
+import com.expressswallows.model.menu.fooditems.pizzas.CheesePizza;
+import com.expressswallows.model.menu.fooditems.pizzas.PepperoniPizza;
+import com.expressswallows.model.menu.fooditems.pizzas.Pizza;
+
 /**
  *
  * @author shahi
  */
-public class PizzaFactory {
-    
+public class PizzaFactory extends AbstractFactory {
+
+
+    @Override
+    public Burger createBurger(String request) {
+        return null;
+    }
+
+    @Override
+    public Pizza createPizza(String request) {
+        if (request.equals("Cheese")) {
+            return new CheesePizza();
+        } else if (request.equals("Pepperoni")) {
+            return new PepperoniPizza();
+        } else if (request.equals("All dressed")) {
+            return new AllDressedPizza();
+        }
+        return null;
+    }
+
+    @Override
+    public Drink createDrink(String request, String size) {
+        return null;
+    }
+
+    @Override
+    public Hotdog createHotdog(String request) {
+        return null;
+    }
+
+    @Override
+    public Fries createFries(String request) {
+        return null;
+    }
 }
