@@ -1,11 +1,26 @@
 package com.expressswallows.utils;
 
 import com.expressswallows.exceptions.*;
+import com.expressswallows.model.menu.factories.AbstractFactory;
+import com.expressswallows.model.menu.factories.FoodFactoryCreator;
+import com.expressswallows.model.menu.fooditems.Food;
+import com.expressswallows.model.restaurant.Order;
+import com.expressswallows.model.restaurant.Payment;
+import com.expressswallows.model.restaurant.Restaurant;
+import com.expressswallows.model.restaurant.users.Address;
+import com.expressswallows.model.restaurant.users.Client;
+import com.expressswallows.model.restaurant.users.Employee;
+
 import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
 import java.sql.*;
+import java.time.LocalDate;
+import java.time.format.DateTimeFormatter;
+import java.util.ArrayList;
+import java.util.Collections;
+import java.util.List;
 
 /**
  * The DatabaseConnectionUtils class is used to interact with the database.
@@ -20,7 +35,7 @@ public class DatabaseConnectionUtils {
     private static final String DATABASE_FILENAME = "FoodDeliveryData.db";
     private static final String DATABASE_URL = "jdbc:sqlite:" + DATABASE_FILENAME;
     private Connection connection;
-    
+
     /**
      * Singleton design pattern. Get the instance of the DatabaseConnectionUtils.
      * 
@@ -37,7 +52,6 @@ public class DatabaseConnectionUtils {
             }
         }
         return instance;
-        
     } 
     
     /**
@@ -112,20 +126,374 @@ public class DatabaseConnectionUtils {
                     + DATABASE_FILENAME + "': " + e.getMessage());
         }
     }
-    
+
+    /**
+     * Get the client with the specified email and password.
+     *
+     * @param email the email of the client to fetch.
+     * @param passwd the password of the client to fetch.
+     * @return the client with the email and password specified. Null if the Client
+     * with such credentials does not exist.
+     */
+    public Client fetchClientWithCredentials(String email, String passwd) {
+        final String SQL = """
+                            SELECT * FROM client WHERE Email = ? AND Password = ?;
+                            """;
+        Client client = null;
+        try (PreparedStatement pstmt = connection.prepareStatement(SQL)) {
+            pstmt.setString(1, email);
+            pstmt.setString(2, passwd);
+
+            try (ResultSet rs = pstmt.executeQuery()) {
+                if (!rs.next()) {
+                    return null;
+                }
+                client = new Client(rs.getString("FirstName"), rs.getString("LastName"),
+                        email, passwd, LocalDate.parse(rs.getString("DateOfBirth")),
+                        rs.getString("PhoneNumber"), fetchAddressById(rs.getInt("AddressID")));
+                // Set the client id
+                client.setClientId(rs.getInt("ClientID"));
+            } catch (SQLException e) {
+                e.printStackTrace();
+            }
+        } catch (SQLException e) {
+            e.printStackTrace();
+        }
+        return client;
+    }
+
+    /**
+     * Get the address by ID.
+     *
+     * @param id the ID of the address.
+     * @return the address object associated with the ID.
+     */
+    private Address fetchAddressById(int id) {
+        final String SQL = """
+                            SELECT * FROM address WHERE id = ?;
+                            """;
+        Address address = null;
+        try (PreparedStatement pstmt = connection.prepareStatement(SQL)) {
+            pstmt.setInt(1, id);
+            try (ResultSet rs = pstmt.executeQuery()) {
+                if (!rs.next()) {
+                    return null;
+                }
+                address = new Address(rs.getString("Street"), rs.getString("StreetNo"),
+                        rs.getString("PostalCode"), Address.City.valueOf(rs.getString("City")));
+            }
+        } catch (SQLException e) {
+            e.printStackTrace();
+        }
+        return address;
+    }
+
+    /**
+     * Get the employee with the specified email and password.
+     *
+     * @param email the email of the employee.
+     * @param passwd the password of the employee.
+     * @return the employee with the specified email and password.
+     */
+    public Employee fetchEmployeeWithCredentials(String email, String passwd) {
+        final String SQL = """
+                            SELECT * FROM employee WHERE Email = ? AND Password = ?;
+                            """;
+        Employee employee = null;
+        try (PreparedStatement pstmt = connection.prepareStatement(SQL)) {
+            pstmt.setString(1, email);
+            pstmt.setString(2, passwd);
+            try (ResultSet rs = pstmt.executeQuery()) {
+                if (!rs.next()) {
+                    return null;
+                }
+                employee = new Employee(rs.getString("FirstName"), rs.getString("LastName"),
+                        email, passwd, LocalDate.parse(rs.getString("DateOfBirth")),
+                        rs.getString("PhoneNumber"));
+                // Set the employeeid
+                employee.setEmployeeId(rs.getInt("EmployeeID"));
+            }
+        } catch (SQLException e) {
+            e.printStackTrace();
+        }
+        return employee;
+    }
+
+    /**
+     * Fetch all restaurant locations.
+     *
+     * @return a list of all restaurant locations
+     */
+    public List<Restaurant> fetchRestaurantLocations() {
+        final String SQL = """
+                            SELECT RestaurantID FROM restaurant;
+                            """;
+        List<Restaurant> restaurants = new ArrayList<>();
+        try (Statement stmt = connection.createStatement()) {
+            try (ResultSet rs = stmt.executeQuery(SQL)) {
+                while (rs.next()) {
+                    int restaurantID = rs.getInt("RestaurantID");
+                    Restaurant restaurant = fetchRestaurantById(restaurantID);
+                    List<Order> orders = fetchRestaurantOrdersInProgress(restaurantID);
+
+                    // Add orders to the restaurant
+                    for (Order order : orders) {
+                        try {
+                            restaurant.addOrder(order);
+                        } catch (InterruptedException e) {
+                            e.printStackTrace();
+                        }
+                    }
+                    // Add to all restaurants
+                    restaurants.add(restaurant);
+                }
+            }
+        } catch (SQLException e) {
+            e.printStackTrace();
+        }
+        return restaurants;
+    }
+
+    /**
+     * Get the restaurant orders that are in progress.
+     * @return
+     */
+    private List<Order> fetchRestaurantOrdersInProgress(int restaurantID) {
+        final String SQL = """
+                            SELECT * FROM order WHERE restaurantID = ?;
+                            """;
+        List<Order> orders = new ArrayList<>();
+        try (PreparedStatement pstmt = connection.prepareStatement(SQL)) {
+            pstmt.setInt(1, restaurantID);
+            try (ResultSet rs = pstmt.executeQuery()) {
+                while (rs.next()) {
+                    if (rs.getString("Status").equals("DELIVERED")) {
+                        continue;
+                    }
+                    // Create the order
+                    Order order = new Order(restaurantID, fetchClientById(rs.getInt("ClientID")));
+                    // Set the order id
+                    order.setOrderId(rs.getInt("OrderID"));
+                    // Add food items to oder
+                    for (Food d: fetchOrderFoods(order.getOrderId())) {
+                        order.addFoodToOrder(d);
+                    }
+                    // Add order to orders
+                    orders.add(order);
+                }
+            }
+        } catch (SQLException e) {
+            e.printStackTrace();
+        }
+        // Sort by oldest to newest
+        Collections.sort(orders, (o1, o2) -> {
+            return o1.getOrderDateTime().compareTo(o2.getOrderDateTime());
+        });
+        return orders;
+    }
+
+    /**
+     * Get a client by his ID.
+     * @param clientId
+     * @return
+     */
+    private Client fetchClientById(int clientId) {
+        final String SQL = """
+                            SELECT * FROM client WHERE ClientID = ?;
+                            """;
+        Client client = null;
+        try (PreparedStatement pstmt = connection.prepareStatement(SQL)) {
+            pstmt.setInt(1, clientId);
+
+            try (ResultSet rs = pstmt.executeQuery()) {
+                if (!rs.next()) {
+                    return null;
+                }
+                client = new Client(rs.getString("FirstName"), rs.getString("LastName"),
+                        rs.getString("Email"), rs.getString("Password"),
+                        LocalDate.parse(rs.getString("DateOfBirth")), rs.getString("PhoneNumber"),
+                        fetchAddressById(rs.getInt("AddressID")));
+                // Set the client id
+                client.setClientId(clientId);
+            } catch (SQLException e) {
+                e.printStackTrace();
+            }
+        } catch (SQLException e) {
+            e.printStackTrace();
+        }
+        return client;
+    }
+
+    /**
+     * Get a restaurant by its ID.
+     * @param restaurantId
+     * @return
+     */
+    private Restaurant fetchRestaurantById(int restaurantId) {
+        final String SQL = """
+                            SELECT * FROM restaurant WHERE restaurantID = ?;
+                            """;
+        Restaurant restaurant = null;
+        try (PreparedStatement pstmt = connection.prepareStatement(SQL)) {
+            pstmt.setInt(1, restaurantId);
+            try (ResultSet rs = pstmt.executeQuery()) {
+                if (!rs.next()) {
+                    return null;
+                }
+                restaurant = new Restaurant(fetchAddressById(rs.getInt("AddressID")), rs.getString("name"),
+                        rs.getDouble("balance"));
+                // TODO: add orders
+                restaurant.setRestaurantId(restaurantId);
+                List<Order> orders = fetchRestaurantOrdersInProgress(restaurantId);
+                for (Order order : orders) {
+                    try {
+                        restaurant.addOrder(order);
+                    } catch (InterruptedException e) {
+                        e.printStackTrace();
+                    }
+                }
+            }
+        } catch (SQLException e) {
+            e.printStackTrace();
+        }
+        return restaurant;
+    }
+
+    private List<Food> fetchOrderFoods(int orderID) {
+        final String SQL = """
+                            SELECT * FROM food f
+                            JOIN orderfood of ON f.FoodID = of.FoodID
+                            WHERE OrderID = ?;
+                            """;
+        List<Food> foods = new ArrayList<>();
+        try (PreparedStatement pstmt = connection.prepareStatement(SQL)) {
+            pstmt.setInt(1, orderID);
+            try (ResultSet rs = pstmt.executeQuery()) {
+                while (rs.next()) {
+                    String[] type = rs.getString("Type").split(":");
+                    FoodFactoryCreator creator = new FoodFactoryCreator();
+                    AbstractFactory factory = creator.getFoodFactory(type[0]);
+                    Food food = null;
+                    switch (type[0]) {
+                        case "Pizza" -> {
+                            food = factory.createPizza(type[1]);
+                        }
+
+                        case "Burger" -> {
+                            food = factory.createBurger(type[1]);
+                        }
+
+                        case "Hotdog" -> {
+                            food = factory.createHotdog(type[1]);
+                        }
+
+                        case "Drink" -> {
+                            food = factory.createDrink(type[1], rs.getString("Size"));
+                        }
+
+                        case "Fries" -> {
+                            food = factory.createFries(type[1]);
+                        }
+                    }
+                    foods.add(food);
+                }
+            }
+        } catch (SQLException e) {
+            e.printStackTrace();
+        }
+        return foods;
+    }
+
+    /**
+     * Get orders by client ID.
+     * @param clientId the client ID of the order.
+     * @return the orders made by the client with the specified client ID.
+     */
+    public List<Order> fetchOrdersByClientId(int clientId) {
+        final String SQL = """
+                           SELECT * FROM order WHERE ClientID = ?;
+                           """;
+        List<Order> orders = new ArrayList<>();
+        try (PreparedStatement pstmt = connection.prepareStatement(SQL)) {
+            pstmt.setInt(1, clientId);
+            try (ResultSet rs = pstmt.executeQuery()) {
+                while (rs.next()) {
+                    // Create the order
+                    Order order = new Order(rs.getInt("RestaurantID"), fetchClientById(clientId));
+                    // Set the order ID
+                    order.setOrderId(rs.getInt("OrderID"));
+                    // Add the foods
+                    List<Food> foods = fetchOrderFoods(order.getOrderId());
+                    for (Food f : foods) {
+                        order.addFoodToOrder(f);
+                    }
+                    // Add order to orders
+                    orders.add(order);
+                }
+            }
+        } catch (SQLException e) {
+            e.printStackTrace();
+        }
+        return orders;
+    }
+
+    /**
+     * Insert an order.
+     * @param order the order to insert
+     */
+    public void insertOrder(Order order) {
+        final String SQL = """
+                            INSERT INTO order (OrderTime, Status, ClientID, RestaurantID)
+                            VALUES (?, ?, ?, ?);
+                            """;
+        try (PreparedStatement pstmt = connection.prepareStatement(SQL)) {
+            DateTimeFormatter formatter = DateTimeFormatter.ofPattern("yyyy-MM-dd HH:mm:ss");
+            pstmt.setString(1, order.getOrderDateTime().format(formatter));
+            pstmt.setString(2, order.getStatus().toString());
+            pstmt.setInt(3, order.getOrderedBy().getClientId());
+            pstmt.setInt(4, order.getRestaurantId());
+            pstmt.executeUpdate();
+            // Get primary key
+            try (ResultSet rs = pstmt.executeQuery()) {
+                if (rs.next()) {
+                    order.setOrderId(rs.getInt(1));
+                }
+            }
+        } catch (SQLException e) {
+            e.printStackTrace();
+        }
+    }
+
+    // TODO
+    private void insertOrderFood(int orderId, int foodId) {
+
+    }
+
+    // TODO
+    public void insertPayment(Payment payment) {
+
+    }
+
+    // TODO
+    public void insertClient(Client client) {
+
+    }
+
+    // TODO: CREATE TRIGGER FOR PAYMENT
+
     /**
      * Create the client table.
      * 
      * @throws DatabaseInitializationException exception thrown when the client
      * table could not be created.
      */
-    private void createClientTable() throws DatabaseInitializationException {
+    private synchronized void createClientTable() throws DatabaseInitializationException {
         final String SQL = """
                      CREATE TABLE IF NOT EXISTS client (
                         ClientID INTEGER PRIMARY KEY,
                         FirstName TEXT NOT NULL,
                         LastName TEXT NOT NULL,
-                        Email TEXT NOT NULL,
+                        Email TEXT UNIQUE NOT NULL,
                         Password TEXT NOT NULL,
                         DateOfBirth TEXT NOT NULL,
                         PhoneNumber TEXT NOT NULL,
@@ -149,7 +517,7 @@ public class DatabaseConnectionUtils {
      * @throws DatabaseInitializationException exception thrown when the address
      * table could not be created.
      */
-    private void createAddressTable() throws DatabaseInitializationException {
+    private synchronized void createAddressTable() throws DatabaseInitializationException {
         final String SQL = """
                     CREATE TABLE IF NOT EXISTS address (
                        AddressID INTEGER PRIMARY KEY,
@@ -174,7 +542,7 @@ public class DatabaseConnectionUtils {
      * @throws DatabaseInitializationException exception thrown when the employee
      * table could not be created.
      */
-    private void createEmployeeTable() throws DatabaseInitializationException {
+    private synchronized void createEmployeeTable() throws DatabaseInitializationException {
         final String SQL = """
                            CREATE TABLE IF NOT EXISTS employee (
                                 EmployeeID INTEGER PRIMARY KEY,
@@ -203,7 +571,7 @@ public class DatabaseConnectionUtils {
      * @throws DatabaseInitializationException exception thrown when the 
      * employee table could not be created.
      */
-    private void createPaymentTable() throws DatabaseInitializationException {
+    private synchronized void createPaymentTable() throws DatabaseInitializationException {
         final String SQL = """
                            CREATE TABLE IF NOT EXISTS payment (
                                 PaymentID INTEGER PRIMARY KEY,
@@ -231,7 +599,7 @@ public class DatabaseConnectionUtils {
      * @throws DatabaseInitializationException exception thrown when the 
      * employee table could not be created.
      */
-    private void createRestaurantTable() throws DatabaseInitializationException {
+    private synchronized void createRestaurantTable() throws DatabaseInitializationException {
         final String SQL = """
                            CREATE TABLE IF NOT EXISTS restaurant (
                                 RestaurantID INTEGER PRIMARY KEY,
@@ -256,7 +624,7 @@ public class DatabaseConnectionUtils {
      * @throws DatabaseInitializationException exception thrown when the 
      * employee table could not be created.
      */
-    private void createOrderTable() throws DatabaseInitializationException {
+    private synchronized void createOrderTable() throws DatabaseInitializationException {
         final String SQL = """
                            CREATE TABLE IF NOT EXISTS "order" (
                                 OrderID INTEGER PRIMARY KEY,
@@ -265,7 +633,7 @@ public class DatabaseConnectionUtils {
                                 ClientID INTEGER NOT NULL,
                                 RestaurantID INTEGER NOT NULL,
                                 CONSTRAINT chk_OrderTime CHECK (OrderTime LIKE '____-__-__ __:__:__'),
-                                CONSTRAINT chk_Status CHECK (Status IN ('Preparing', 'Delivering', 'Completed')),
+                                CONSTRAINT chk_Status CHECK (Status IN ('IN_PROGRESS', 'DELIVERING', 'DELIVERED')),
                                 CONSTRAINT fk_ClientID FOREIGN KEY (ClientID) REFERENCES client(ClientID),
                                 CONSTRAINT fk_RestaurantID FOREIGN KEY (RestaurantID) REFERENCES restaurant(RestaurantID)
                            );
@@ -284,7 +652,7 @@ public class DatabaseConnectionUtils {
      * @throws DatabaseInitializationException exception thrown when the 
      * employee table could not be created.
      */
-    private void createOrderFoodTable() throws DatabaseInitializationException {
+    private synchronized void createOrderFoodTable() throws DatabaseInitializationException {
         final String SQL = """
                            CREATE TABLE IF NOT EXISTS orderfood (
                                 OrderID INTEGER NOT NULL,
@@ -307,7 +675,7 @@ public class DatabaseConnectionUtils {
      * @throws DatabaseInitializationException exception thrown when the 
      * employee table could not be created.
      */
-    private void createFoodTable() throws DatabaseInitializationException {
+    private synchronized void createFoodTable() throws DatabaseInitializationException {
         final String SQL = """
                            CREATE TABLE IF NOT EXISTS food (
                                 FoodID INTEGER PRIMARY KEY,
@@ -341,7 +709,7 @@ public class DatabaseConnectionUtils {
      * @throws DatabaseInitializationException exception thrown when the 
      * employee table could not be created.
      */
-    private void createFoodToppingTable() throws DatabaseInitializationException {
+    private synchronized void createFoodToppingTable() throws DatabaseInitializationException {
         final String SQL = """
                            CREATE TABLE IF NOT EXISTS foodtopping (
                                 ToppingID INTEGER NOT NULL,
@@ -364,7 +732,7 @@ public class DatabaseConnectionUtils {
      * @throws DatabaseInitializationException exception thrown when the 
      * employee table could not be created.
      */
-    private void createToppingTable() throws DatabaseInitializationException {
+    private synchronized void createToppingTable() throws DatabaseInitializationException {
         final String SQL = """
                            CREATE TABLE IF NOT EXISTS topping (
                                 ToppingID INTEGER PRIMARY KEY,
@@ -386,7 +754,7 @@ public class DatabaseConnectionUtils {
      * @throws SQLException exception thrown when the statement fails to 
      * execute. 
      */
-    private void executeSQL(String sql) throws SQLException {
+    private synchronized void executeSQL(String sql) throws SQLException {
         // Create the statement
         try (Statement stmt = connection.createStatement()) {
             // Execute the SQL
