@@ -4,17 +4,35 @@
  */
 package com.expressswallows.view;
 
+import com.expressswallows.model.menu.fooditems.Food;
+import com.expressswallows.model.restaurant.Order;
+import com.expressswallows.model.restaurant.users.Client;
+import com.expressswallows.utils.Utils;
+import java.util.ArrayList;
+
+import java.util.List;
+import java.util.ResourceBundle;
+import javax.swing.DefaultListModel;
+
 /**
  *
  * @author shahi
  */
 public class FormViewCart extends javax.swing.JFrame {
 
+    Order order;
+    Client client;
+    List<Food> foods = order.getFoods();
+
     /**
      * Creates new form FormViewCart
      */
-    public FormViewCart() {
+    public FormViewCart(Client client, Order order) {
         initComponents();
+        this.order = order;
+        this.client = client;
+        update();
+        loadOrder(order.getFoods());
     }
 
     /**
@@ -29,6 +47,8 @@ public class FormViewCart extends javax.swing.JFrame {
         viewcartLbl = new javax.swing.JLabel();
         backBtn = new javax.swing.JButton();
         payBtn = new javax.swing.JButton();
+        jScrollPane1 = new javax.swing.JScrollPane();
+        listOrders = new javax.swing.JList<>();
 
         setDefaultCloseOperation(javax.swing.WindowConstants.EXIT_ON_CLOSE);
 
@@ -49,28 +69,35 @@ public class FormViewCart extends javax.swing.JFrame {
             }
         });
 
+        jScrollPane1.setViewportView(listOrders);
+
         javax.swing.GroupLayout layout = new javax.swing.GroupLayout(getContentPane());
         getContentPane().setLayout(layout);
         layout.setHorizontalGroup(
             layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
             .addGroup(layout.createSequentialGroup()
-                .addGroup(layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-                    .addGroup(layout.createSequentialGroup()
-                        .addGap(141, 141, 141)
-                        .addComponent(viewcartLbl))
-                    .addGroup(layout.createSequentialGroup()
-                        .addGap(123, 123, 123)
-                        .addComponent(backBtn)
-                        .addGap(18, 18, 18)
-                        .addComponent(payBtn)))
+                .addGap(123, 123, 123)
+                .addComponent(backBtn)
+                .addGap(18, 18, 18)
+                .addComponent(payBtn)
                 .addContainerGap(115, Short.MAX_VALUE))
+            .addGroup(javax.swing.GroupLayout.Alignment.TRAILING, layout.createSequentialGroup()
+                .addContainerGap(javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
+                .addComponent(jScrollPane1, javax.swing.GroupLayout.PREFERRED_SIZE, 229, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addContainerGap(javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE))
+            .addGroup(layout.createSequentialGroup()
+                .addGap(125, 125, 125)
+                .addComponent(viewcartLbl)
+                .addContainerGap(javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE))
         );
         layout.setVerticalGroup(
             layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
             .addGroup(layout.createSequentialGroup()
                 .addContainerGap()
                 .addComponent(viewcartLbl)
-                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED, 207, Short.MAX_VALUE)
+                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
+                .addComponent(jScrollPane1, javax.swing.GroupLayout.DEFAULT_SIZE, 183, Short.MAX_VALUE)
+                .addGap(18, 18, 18)
                 .addGroup(layout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
                     .addComponent(backBtn)
                     .addComponent(payBtn))
@@ -80,53 +107,35 @@ public class FormViewCart extends javax.swing.JFrame {
         pack();
     }// </editor-fold>//GEN-END:initComponents
 
+    private void update() {
+        ResourceBundle rb = ResourceBundle.getBundle("messages", Utils.currentLocale);
+        viewcartLbl.setText(rb.getString("yourcart"));
+        payBtn.setText(rb.getString("pay"));
+        backBtn.setText(rb.getString("back"));
+    }
+
     private void payBtnActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_payBtnActionPerformed
         dispose();
-        new FormPayment().setVisible(true);
+        new FormPayment(client, order).setVisible(true);
     }//GEN-LAST:event_payBtnActionPerformed
 
     private void backBtnActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_backBtnActionPerformed
         dispose();
-        new FormClientMainMenu().setVisible(true);
+        new FormClientMainMenu(client, order).setVisible(true);
     }//GEN-LAST:event_backBtnActionPerformed
 
-    /**
-     * @param args the command line arguments
-     */
-    public static void main(String args[]) {
-        /* Set the Nimbus look and feel */
-        //<editor-fold defaultstate="collapsed" desc=" Look and feel setting code (optional) ">
-        /* If Nimbus (introduced in Java SE 6) is not available, stay with the default look and feel.
-         * For details see http://download.oracle.com/javase/tutorial/uiswing/lookandfeel/plaf.html 
-         */
-        try {
-            for (javax.swing.UIManager.LookAndFeelInfo info : javax.swing.UIManager.getInstalledLookAndFeels()) {
-                if ("Nimbus".equals(info.getName())) {
-                    javax.swing.UIManager.setLookAndFeel(info.getClassName());
-                    break;
-                }
-            }
-        } catch (ClassNotFoundException ex) {
-            java.util.logging.Logger.getLogger(FormViewCart.class.getName()).log(java.util.logging.Level.SEVERE, null, ex);
-        } catch (InstantiationException ex) {
-            java.util.logging.Logger.getLogger(FormViewCart.class.getName()).log(java.util.logging.Level.SEVERE, null, ex);
-        } catch (IllegalAccessException ex) {
-            java.util.logging.Logger.getLogger(FormViewCart.class.getName()).log(java.util.logging.Level.SEVERE, null, ex);
-        } catch (javax.swing.UnsupportedLookAndFeelException ex) {
-            java.util.logging.Logger.getLogger(FormViewCart.class.getName()).log(java.util.logging.Level.SEVERE, null, ex);
+    private void loadOrder(List<Food> foods) {
+        DefaultListModel<String> listModel = new DefaultListModel<>();
+        for (Food food : foods) {
+            listModel.addElement(food.toString());
         }
-        //</editor-fold>
-
-        /* Create and display the form */
-        java.awt.EventQueue.invokeLater(new Runnable() {
-            public void run() {
-                new FormViewCart().setVisible(true);
-            }
-        });
+        listOrders.setModel(listModel);
     }
-
+    
     // Variables declaration - do not modify//GEN-BEGIN:variables
     private javax.swing.JButton backBtn;
+    private javax.swing.JScrollPane jScrollPane1;
+    private javax.swing.JList<String> listOrders;
     private javax.swing.JButton payBtn;
     private javax.swing.JLabel viewcartLbl;
     // End of variables declaration//GEN-END:variables
